break out ts -> jsonschema, export it
