+++ transformers/typescript/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./ts-prompt-transformer";
+export * from "./ts-type-to-json-schema";
 import transformer from "./ts-prompt-transformer";
 export default transformer;
